requirements.txt
anyascii==0.3.2
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
boto3==1.28.8
botocore==1.31.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.5
cryptography==41.0.2
cssselect2==0.7.0
Django==4.1.8
django-cockroachdb==4.1.1
django-environ==0.10.0
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-ses==3.5.0
django-sso==3.0.2
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
l18n==2021.3
lxml==4.9.3
numpy==1.24.2
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.0.0
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf==3.12.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
s3transfer==0.6.1
sentry-sdk==1.28.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
svglib==1.5.1
telepath==0.3
tinycss2==1.2.1
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.15
wagtail==4.2.2
webencodings==0.5.1
Willow==1.4.1
xhtml2pdf==0.2.11
